+++ Changes.md
@@ -1,6 +1,19 @@
 
+#
+### **+05:30 11:40:31 PM 13-08-2023, Sunday**
+
+  * Added more debug messages.
+  * Added `isCRCSet` to `calculateCRC()` to tell if the CRC is set or not.
+  * Added `print()' to `CSE_ModbusRTU_ADU` to print the ADU as a hex string.
+  * Added receive timeout to `CSE_ModbusRTU:: receive`.
+  * Fixed issues with `CSE_ModbusRTU:: send` by adding begin and end transactions.
+  * Added `CSE_ModbusRTU_Client` class.
+  * Examples updated and working.
+  * New version **0.0.3**.
+
+#
 ### **+05:30 08:41:17 PM 02-08-2023, Wednesday**
 
   * Added more documentation.
-  * New version 0.0.2.
+  * New version **0.0.2**.